+++ CHANGELOG.md
@@ -28,6 +28,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Converted `<manifold-data-deprovision-button>` to use GraphQL (#604)
 - Converted `<manifold-plan-cost>` to GraphQL (#605)
 - Updated Stencil to v1.6.1 (#606)
+- Owner ID no longer fetched to create/delete resources (#608)
 
 ### Breaking Changes: